+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## WIP
+## v1.3 (2021-04-19)
 
 - Adds a `case_sensitive` option to `test()` and `match()` function. (`True` by default)
 - Allows editing and manual specification of `Matcher.regex` and `Matcher.converters`.